Enhance authentication pages with user data props and dropdown navigation; update routing for StatusCV and CompanyProfile
