+++ resources/js/Pages/Services.vue
@@ -1,7 +1,14 @@
 <script setup>
-import { Head } from '@inertiajs/vue3';
-import { Link } from '@inertiajs/vue3';
+import { Head, Link, router } from '@inertiajs/vue3';
+import Dropdown from '@/Components/Dropdown.vue';
+import DropdownLink from '@/Components/DropdownLink.vue';
+import { ref } from 'vue';
 
+const showingNavigationDropdown = ref(false);
+
+const logout = () => {
+    router.post(route('logout'));
+};
 
 // Jika kita perlu menggunakan asset dari Laravel, kita bisa membuat computed property
 const asset = (path) => `/assets/${path}`;
@@ -19,7 +26,7 @@ const asset = (path) => `/assets/${path}`;
             <nav class="space-x-10 -ml-60 mr-80">
                 <Link 
                     class="text-blue-900 transition-colors duration-200 hover:text-[#4D62D7]" 
-                    href="/companyprofile"
+                    href="/companyProfile"
                 >
                     Tentang Kami
                 </Link>
@@ -45,18 +52,44 @@ const asset = (path) => `/assets/${path}`;
             <div class="-mr-72">
                 <Link 
                     class="text-[#5932EA] font-bold transition-colors duration-200 hover:text-[#4D62D7]" 
-                    href="/uploadcv"
+                    href="/hiring"
                 >
                     We are hiring!!
                 </Link>
             </div>
             <div class="flex items-center space-x-2">
-                <Link href="/login">
-                    <button class="bg-green-200 text-green-800 px-4 py-2 rounded mr-3">
-                        Log In
-                    </button>
-                </Link>
-                <i class="fas fa-user-circle text-blue-900 text-3xl"></i>
+                <template v-if="$page.props.auth.user">
+                    <Dropdown align="right" width="48">
+                        <template #trigger>
+                            <span class="inline-flex rounded-md">
+                                <button
+                                    type="button"
+                                    class="inline-flex items-center px-3 py-2 text-sm leading-4 font-medium rounded-md text-gray-500 bg-white hover:text-gray-700 focus:outline-none transition ease-in-out duration-150"
+                                >
+                                    {{ $page.props.auth.user.name }}
+                                    <i class="fas fa-user-circle text-blue-900 text-3xl ml-2"></i>
+                                </button>
+                            </span>
+                        </template>
+
+                        <template #content>
+                            <DropdownLink :href="route('profile.edit')">
+                                Profile
+                            </DropdownLink>
+                            <DropdownLink :href="route('logout')" method="post" as="button">
+                                Log Out
+                            </DropdownLink>
+                        </template>
+                    </Dropdown>
+                </template>
+                <template v-else>
+                    <Link href="/login">
+                        <button class="bg-green-200 text-green-800 px-4 py-2 rounded mr-3">
+                            Log In
+                        </button>
+                    </Link>
+                    <i class="fas fa-user-circle text-blue-900 text-3xl"></i>
+                </template>
             </div>
         </header>
 
@@ -65,7 +98,7 @@ const asset = (path) => `/assets/${path}`;
             <div class="bg-yellow-100 p-4 flex justify-center space-x-16 w-2/5 mx-auto">
                 <Link 
                     class="text-gray-700 text-center transition-colors duration-200 hover:text-[#CDC052]" 
-                    href="/companyprofile"
+                    href="/companyProfile"
                 >
                     Company Profile
                 </Link>
