+++ resources/js/Pages/StatusCV.vue
@@ -1,9 +1,21 @@
-# UploadConfirmation.vue
 <script setup>
 import { Head, Link } from '@inertiajs/vue3';
+import Dropdown from '@/Components/Dropdown.vue';
+import DropdownLink from '@/Components/DropdownLink.vue';
+import { ref } from 'vue';
+
+const showingNavigationDropdown = ref(false);
+
+const logout = () => {
+    router.post(route('logout'));
+};
 
 // Define props for user data
 defineProps({
+    auth: {
+        type: Object,
+        required: true,
+    },
     userData: {
         type: Object,
         default: () => ({
@@ -31,7 +43,7 @@ defineProps({
             <nav class="space-x-10 -ml-60 mr-80">
                 <Link 
                     class="text-blue-900 transition-colors duration-200 hover:text-[#4D62D7]" 
-                    href="/companyprofile"
+                    href="/companyProfile"
                 >
                     Tentang Kami
                 </Link>
@@ -63,12 +75,38 @@ defineProps({
                 </Link>
             </div>
             <div class="flex items-center space-x-2">
-                <Link href="/login">
-                    <button class="bg-green-200 text-green-800 px-4 py-2 rounded mr-3">
-                        Log In
-                    </button>
-                </Link>
-                <i class="fas fa-user-circle text-blue-900 text-3xl"></i>
+                <template v-if="$page.props.auth.user">
+                    <Dropdown align="right" width="48">
+                        <template #trigger>
+                            <span class="inline-flex rounded-md">
+                                <button
+                                    type="button"
+                                    class="inline-flex items-center px-3 py-2 text-sm leading-4 font-medium rounded-md text-gray-500 bg-white hover:text-gray-700 focus:outline-none transition ease-in-out duration-150"
+                                >
+                                    {{ $page.props.auth.user.name }}
+                                    <i class="fas fa-user-circle text-blue-900 text-3xl ml-2"></i>
+                                </button>
+                            </span>
+                        </template>
+
+                        <template #content>
+                            <DropdownLink :href="route('profile.edit')">
+                                Profile
+                            </DropdownLink>
+                            <DropdownLink :href="route('logout')" method="post" as="button">
+                                Log Out
+                            </DropdownLink>
+                        </template>
+                    </Dropdown>
+                </template>
+                <template v-else>
+                    <Link href="/login">
+                        <button class="bg-green-200 text-green-800 px-4 py-2 rounded mr-3">
+                            Log In
+                        </button>
+                    </Link>
+                    <i class="fas fa-user-circle text-blue-900 text-3xl"></i>
+                </template>
             </div>
         </header>
